requirements.txt
torch>=2.4.0
torchvision>=0.19.0
diffusers>=0.31.0
transformers>=4.40.0
accelerate>=0.28.0
datasets>=2.18.0
wandb>=0.16.0
opencv-python>=4.9.0
pillow>=10.2.0
numpy>=1.26.0
tqdm>=4.66.0
omegaconf>=2.3.0
einops>=0.7.0
xformers>=0.0.25
bitsandbytes>=0.43.0
peft>=0.10.0
safetensors>=0.4.2
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
av>=11.0.0
decord>=0.6.0
scikit-image>=0.22.0
matplotlib>=3.8.0
seaborn>=0.13.0
tensorboard>=2.16.0
huggingface-hub>=0.21.0
timm>=0.9.16
ftfy>=6.1.3
regex>=2023.12.25

# Additional dependencies for unified training
python-dotenv>=1.0.0
psutil>=5.9.0
scipy>=1.11.0
requests>=2.31.0
urllib3>=2.0.0
clip-by-openai>=1.0.1
salesforce-lavis>=1.0.2
pathlib2>=2.3.7
typing-extensions>=4.8.0
packaging>=23.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
joblib>=1.3.0
rich>=13.0.0
click>=8.1.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
nvidia-ml-py>=12.535.0
gpustat>=1.1.0
aiohttp>=3.9.0
diskcache>=5.6.0
lz4>=4.3.0
zstandard>=0.22.0
librosa>=0.10.0
soundfile>=0.12.0
pyyaml>=6.0
pandas>=2.0.0

# Flask Web Interface
flask>=3.0.0
flask-socketio>=5.3.0
werkzeug>=3.0.0
